Added bankIntegrationsDAO to fetch banks added from accounting softwares

+++ nanopay/src/net/nanopay/integration/User.js
@@ -3,7 +3,7 @@ foam.CLASS({
 
   properties: [
     {
-      class: 'Long',
+      class: 'Int',
       name: 'integrationCode',
       value: -1,
       hidden: true,
